requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
openai>=1.12.0
python-dotenv>=1.0.0
